requirements.txt
pandas
scikit-learn
numpy
matplotlib
flask
python-box # config box
pyYAML # Read YAML file
ensure
joblib
types-PyYAML
Flask-Cors
joblib
seaborn
pymongo
streamlit
pip-chill==1.0.1
LightGBM
category_encoders
requests
Flask-Mail
mailtrap
psycopg2

# mlflow 
mlflow
dagshub
dvc

-e .